+++ src/main/java/cc/uncarbon/interceptor/AdminSaTokenParseInterceptor.java
@@ -5,7 +5,7 @@ import cc.uncarbon.framework.core.context.TenantContextHolder;
 import cc.uncarbon.framework.core.context.UserContext;
 import cc.uncarbon.framework.core.context.UserContextHolder;
 import cc.uncarbon.framework.web.util.IPUtil;
-import cc.uncarbon.module.sys.util.AdminStpUtil;
+import cc.uncarbon.module.adminapi.util.AdminStpUtil;
 import cn.dev33.satoken.session.SaSession;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
